# swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "version": "1.0.0",
      "title": "Plainx middleware api documentation",
      "description": "API documentation for plainx middleware",
      "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "host": "https://plainx-midddleware-engine.azurewebsites.net",
    "basePath": "/",
    "components": {
        "securitySchemes": {
          "jwt": {
            "type": "http",
            "scheme": "bearer",
            "name": "x-access-token",
            "in": "header",
            "bearerFormat": "JWT"
          }
        }
      },
      "security": [{
        "jwt": []
      }],
...
    "schemes": [
      "https"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
...
      "/model/{modelId}": {
        "parameters": [
          {
            "name": "modelId",
            "in": "path",
            "required": true,
            "description": "ID of model that we want to find",
            "type": "string"
          }
        ],
        "get": {
          "tags": [
            "Models"
          ],
          "summary": "Get model with given ID",
          "responses": {
            "200": {
              "description": "Model is found",
              "schema": {
                "$ref": "#/definitions/Model"
              }
            },
            "403": {
              "description": "Failed to authenticate token."
            }
          }
        }
      },
      "/model/local_explainer/{projectId}":{
        "parameters":[
         {
           "name":"projectId",
           "in": "path",
           "required": true,
           "description": "ID of project that we want to find details",
           "type": "string"
         },
         {
            "in": "query",
            "name": "claimNumber",
            "schema":
            {
            "type": "integer",
            "description": "The Claim Number for which we want to find details"
            }
         }
        ],
        "get": {
         "tags": [
           "Models"
         ],
         "summary": "Get Local Explainer details of project with given ID",
           "responses": {
             "200": {
               "description": "Local Explainer details found successfully",
               "schema": {
                 "$ref": "#/definitions/Project"
               }
             },
             "403": {
               "description": "Failed to authenticate token."
             },
             "404":{
               "description": "No project found with given id"
             },
             "500":{
               "description":"There was a problem fetching data.Please try again after sometime"
             }
           }
         }
       },
       "/model/local_explainer/records/{projectId}":{
         "parameters":[
          {
            "name":"projectId",
            "in": "path",
            "required": true,
            "description": "ID of project that we want to find details",
            "type": "string"
          }
         ],
         "get": {
          "tags": [
            "Models"
          ],
          "summary": "Get Local Explainer Records details of project with given ID",
            "responses": {
              "200": {
                "description": "Local Explainer  Records details found successfully",
                "schema": {
                  "$ref": "#/definitions/Project"
                }
              },
              "403": {
                "description": "Failed to authenticate token."
              },
              "404":{
                "description": "No project found with given id"
              },
              "500":{
                "description":"There was a problem fetching data.Please try again after sometime"
              }
            }
          }
        },
      "/notification/all/{userId}": {
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "description": "ID of user whose notifications we want to get",
            "type": "string"
          }
        ],
        "get": {
          "tags": [
            "Notifications"
          ],
          "summary": "Get all notifications for user",
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/Notifications"
              }
            },
            "403": {
              "description": "Failed to authenticate token."
            }
          }
        }
      },
...
      "/notification/{userId}": {
        "post": {
          "tags": [
            "Notifications"
          ],
          "description": "Create a notification for user",
          "parameters": [
            {
              "name": "userId",
              "in": "path",
              "required": true,
              "description": "ID of user who we want to send notification to",
              "type": "string"
            },
            {
              "name": "Notification",
              "in": "body",
              "description": "Notification that we want to create",
              "schema": {
                "type": "object",
                "properties": {
                  "message": { 
                    "description": "Message",
                    "type": "string"
                  },
                  "type": {
                    "description": "Type",
                    "type": "string"
                  }
                },
                "required": ["message", "type"] 
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { 
                      "description": "Message",
                      "type": "string"
                    },
                    "type": {
                      "description": "Type",
                      "type": "string"
                    }
                  },
                  "required": ["message", "type"] 
                }
              }
            }
          },
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "The details of the newly created notification."
            },
            "403": {
              "description": "Failed to authenticate token."
            }
          }
        }
      },
      "/notification/{notificationId}": {
        "parameters": [
          {
            "name": "notificationId",
            "in": "path",
            "required": true,
            "description": "ID of notification that we want to update",
            "type": "string"
          }
        ],
        "put": {
          "tags": [
            "Notifications"
          ],
          "summary": "Update notification with given notification ID",
          "responses": {
            "200": {
              "description": "Notification was updated to isRead = true.",
              "schema": {
                "$ref": "#/definitions/Notification"
              }
            },
            "403": {
              "description": "Failed to authenticate token."
            }
          }
        }
      }
    },
    "definitions": {
...
      "Project": {
        "properties": {
          "userId": { "type": "string"},
          "projectName": { "type": "string" },
          "projectStage": { "type": "string" },
          "industry" : { "type": "string" },
          "useCase": { "type": "string" },
          "model": {"type": "object", "$ref": "#/definitions/Model"},
          "datasetFile": { "type": "string" },
          "datasetSize": { "type": "string" },
          "datasetName": { "type": "string" }
        },
        "required": [
            "userId"
        ]
      },
      "Model": {
        "properties": {
          "modelName": { "type": "string"},
          "modelType": { "type": "string"},
          "modelFile": { "type": "string"},
          "dataToPredict": { "type": "string"}
        }
      },
...
      "Notifications":{
        "properties": {
          "userId": { "type": "string"},
          "notificationId": { "type": "string"},
          "type": { "type": "string"},
          "message": { "type": "string"},
          "dateCreated": { "type": "string"},
          "isRead": { "type": "boolean"}
        },
        "required": [
            "userId"
        ]
        },
...
    }
  }